Extract users table filters to a separate component
