+++ src/components/Screen/UsersScreen/FiltersToolbar/style.module.scss
@@ -1,0 +1,23 @@
+.filters-toolbar {
+  width: 100%;
+
+  .filter-wrapper {
+    flex-wrap: nowrap;
+    margin-right: 40px;
+
+    label {
+      color: var(--neutral900);
+      font-family: var(--font-primary);
+      white-space: nowrap;
+      margin-right: 10px;
+    }
+
+    .select-filter {
+      width: 145px;
+
+      &.multi {
+        width: 180px;
+      }
+    }
+  }
+}
